requirements.txt
Flask-Cors==4.0.0
Jinja2==3.1.2
MarkupSafe==2.1.3
bcrypt==4.0.1
blinker==1.7.0
click==8.1.7
colorama==0.4.6
flask==3.0.0
importlib-metadata==6.8.0
itsdangerous==2.1.2
mysql-connector-python==8.2.0
pip==22.0.4
protobuf==4.21.12
setuptools==58.1.0
werkzeug==3.0.1
zipp==3.17.0

